requirements.txt
Flask
faiss-cpu
PyMuPDF
sentence-transformers
bert-score
google-generativeai
